+++ packages/multichain-account-service/src/tests/accounts.ts
@@ -214,8 +214,6 @@ export const MOCK_BTC_P2WPKH_ACCOUNT_1: Bip44Account<InternalAccount> = {
     },
     snap: {
       id: 'mock-btc-snap-id',
-      enabled: true,
-      name: 'Mock Bitcoin Snap',
     },
   },
 };
@@ -243,8 +241,6 @@ export const MOCK_BTC_P2TR_ACCOUNT_1: Bip44Account<InternalAccount> = {
     },
     snap: {
       id: 'mock-btc-snap-id',
-      enabled: true,
-      name: 'Mock Bitcoin Snap',
     },
   },
 };
@@ -353,9 +349,7 @@ export class MockAccountBuilder<Account extends KeyringAccount> {
   withSnapId(snapId: SnapId) {
     if (isInternalAccount(this.#account)) {
       this.#account.metadata.snap = {
-        enabled: true,
         id: snapId,
-        name: `Name: ${snapId}`,
       };
     }
     return this;